Title: 특정 문자로 시작하는 문자열, Time: 143ms, Memory: 10MB, Status: Passed - CodeTree

+++ 240309/특정 문자로 시작하는 문자열/strings-that-start-with-a-specific-character.java
@@ -25,7 +25,7 @@ public class Main {
             }
         }
 
-        double avg = sum/cnt;
+        double avg = (double) sum/cnt;
 
         System.out.printf("%d %.2f",cnt,avg);
